Id to number, parsing env with zod

+++ backend/db/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS
 CREATE TABLE IF NOT EXISTS
    tasks (
       id INTEGER AUTO_INCREMENT PRIMARY KEY NOT NULL,
-      project_id INTEGER NOT NULL,
+      project_id BIGINT NOT NULL,
       `name` VARCHAR(255) NOT NULL,
       completed BOOLEAN NOT NULL,
       priority ENUM('none', 'low', 'medium', 'high') NOT NULL,
